Extract it into a function so we can test it
