+++ app/src/lib/stores/index.ts
@@ -9,3 +9,4 @@ export * from './repositories-store'
 export * from './sign-in-store'
 export * from './token-store'
 export * from './pull-request-store'
+export { UpstreamRemoteName } from './helpers/find-upstream-remote'
